+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+
+### Features
+
+* Reading from files now emits the include directives
+* `BeancountFile::read_files` to read files directly into a `BeancountFile`
+
+
 ### Dependencies
 
 * Remove dependency on `thiserror`